Rationalize error strings using constants
Error constructors now use constants for error strings, or in few cases to build error strings, to avoid an excessive use of string literals in the code. Most error strings now do not reference the context that generated the error, as this type of reporting is a specific task left to logging.