Fix DBus method call in conditions
The method call has been split in two cases: with or without arguments. Also, the argument list has been unwrapped from `Value`: when wrapped, it confused DBus providing a wrong call signature, so the actual method to be called was not found. Minor documentation fixes and version bump in order to verify that the correct version is used. Fixes #7.