INTERMEDIATE: Control event service management (Step 1)
First attempts at bringing event listening services management to the event registry instead of the main function, and to dynamically handle event listening services threads. This involves strong modifications to both the event registry code and to the event listening service tasks.
WARNING: this code is *NOT* working, nor does it compile, it is committed as an intermediate step in order to set the development direction.