INTERMEDIATE: Event service management (Step 3)
Synchronization issues in the event service management thread have been fixed, and now the event listening services seem to be properly installed. No dynamic service management has been tried yet though. The produced executable seems to work almost exactly as the one without event service manager.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.