INTERMEDIATE: Dynamic add/remove/replace items: conditions
Add API and synchronization tools to handle items (add, remove, and replace) in a running condition registry. Modifications are deferred when the registry is busy running tasks.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.