INTERMEDIATE: API to get config parameters (pt. 1)
Create an API to simplify retrieval of common configuration parameters in an easier/cleaner way, so that it can be used both to actually configure items and to check configuration for correctness without actually creating the item.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.