INTERMEDIATE: API to get config parameters (pt. 2)
Create an API to simplify retrieval of common configuration parameters in an easier/cleaner way, so that it can be used both to actually configure items and to check configuration for correctness without actually creating the item: this second step adds support for lists of simple homogeneous values (booleans, integers, floats, and strings) including checks - both generic and common. The new API is used, at the moment, only in two types of item: time based conditions and command based tasks. All items will eventually use the API.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.