INTERMEDIATE: API to get config parameters (pt. 3)
The new API is now used in all existing items: code is actually simplified and the two task items have been provided with config-check-only functions that will be used to check configurations without interference with the running instance. The application is still able to load its configuration at startup and run exactly as before. Some error messages have however been suppressed in favor of more generic ones.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.