INTERMEDIATE: item configuration checks
Now every item provides an interface to check correctness of a provided configuration chunk without actually loading the configuration into a new item instance.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.