INTERMEDIATE: full configuration check
A global check function has been provided, that checks a whole configuration file without alteration of the current status, in order to decide whether or not a new configuration can be implemented - and thus would be loaded into a running instance.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.