INTERMEDIATE: Implement the configure command
Finally the helpers to reconfigure a running instance have been implemented and their combination has been associated to a `configure` command which can be passed through the standard input, along with the name of the file to read the new configuration from. The results are promising: the first tests only change modified items and delete removed ones as expected. Now more thorough tests have to be performed, maybe some code checks (eg. consistency of log messages) and optimizations, and documentation updates. After that this branch will be merged into main. NOTE: the reconfiguration process *does not* alter global parameters: this would involve actually stopping the scheduler and restarting it, which is not easy for the moment.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.