INTERMEDIATE: Change event synchronization logic
The synchronization logic in the event registry had to be completely changed in order to avoid mutual lock of event items while checking for their presence in the registry: since events never change status during their life, a more permissive RwLock based logic has been implemented.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.