INTERMEDIATE: Change event listener logic
In order to allow for event listeners to be modified at runtime, the underlying logic had to be completely changed, to the point that event listeners themselves now are built exploiting async protocols in order to also listen for a `quit` command sent by the registry when an event has to be removed or replaced. This technique still has to be implemented for DBus based events (it is only required for listeners that use a dedicated thread) and has to be improved for FSChange events too.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.