INTERMEDIATE: Change event listener logic
Use the same queue that the notify crate uses to enqueue the "quit" signal, and wait for it in a separate lightweight thread that feeds the same future: this allows for a single future to need handling and a more straightforward way to react to events. A `FilesystemChangeEvent` is now able to stop its listener when a newly submitted configuration does not contain it. DBus events still need to be implemented in this way, and code cleanup might be required.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.