INTERMEDIATE: Change event listener logic
Apply the same technique used in #2b59b54dedded8674ab70b537b2b16fac32d6ef6 to the DBus listener, to unify the queue for target and quit messages. Also use assertions for code that should panic at debug time.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.