INTERMEDIATE: Change event listener logic (FIX)
Fix a lock condition that prevented an event listener to be removed from the registry by the main event service manager: the internal running status now is reset when the service stops listening for events.
WARNING: this code is *NOT* properly tested and might not be working, it is committed as an intermediate step in order to set the development direction.