Implement #41 and #40
This commit enables retries for non recurring conditions whose tasks fail, and debouncing of successful condition checks for conditions that do not depend on time or events. The features are implemented via new parameters, the omission of which just reproduces the behavior of previous versions. Still to be tested on Linux for DBus method conditions, and documentation has to be reviewed before a pull request. The changes seemed worth a revision number bump.