Fix conditions not reloading
Fix bug #45 by correctly checking whether or not a condition of certain types (command, lua, dbus, event) is identical to the one already present with the same name. Also, log a message when an item is reconfigured, not only when the configuration is skipped. Version bumped for a new release.