Fix mutual locks in condition tick
After a condition tick/check/run, the check for presence in the reset queue (or in the suspend queue) and request to reset (or suspend respectively) caused a mutual lock due to overzealous use of the external registry interface also in a core function such as `tick`: this has been changed to avoid possibly blocking calls, and now the scheduler works as expected. Also, some utility function have been split to avoid redundancy and some symbol renaming has taken place. The change is scheduled to reach the main branch by next week.